+++ db/sqlc/entry.sql.go
@@ -14,8 +14,8 @@ INSERT INTO entries (account_id, amount) VALUES ($1,$2) RETURNING id, account_id
 `
 
 type CreateEntryParams struct {
-	AccountID int64 `json:"account_id"`
-	Amount    int64 `json:"amount"`
+	AccountID int64   `json:"account_id"`
+	Amount    float64 `json:"amount"`
 }
 
 func (q *Queries) CreateEntry(ctx context.Context, arg CreateEntryParams) (Entry, error) {
