+++ util/random.go
@@ -33,8 +33,8 @@ func RandomOwner() string {
 }
 
 // RandomMoney generates a random amount of money
-func RandomMoney() int64 {
-	return RandomInt(0, 1000)
+func RandomMoney() float64 {
+	return float64(RandomInt(0, 1000))
 }
 
 // RandomCurrency generates a random currency
